Move cache.go to engine submodule
