+++ engine/cache.go
@@ -1,4 +1,4 @@
-package cache
+package engine
 
 type CachedEval struct {
 	Hash  uint64
