+++ engine/game.go
@@ -2,8 +2,6 @@ package engine
 
 import (
 	"fmt"
-
-	. "github.com/amanjpro/zahak/cache"
 )
 
 type Game struct {
