+++ search/types.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"time"
 
-	. "github.com/amanjpro/zahak/cache"
 	. "github.com/amanjpro/zahak/engine"
 )
